fix: mypy strict temizliği (100→0) + schema-drift guard 12-kolon (OOD) + U-009
CI Lint geçtikten sonra ilk kez koşan 2 pre-existing job düzeltildi:

1) mypy (mypy.ini disallow_untyped_defs/warn_return_any strict): 100 hata / 40 dosya → 0
   ("Success: no issues found in 39 source files"). Tüm düzeltmeler RUNTIME-NÖTR:
   - no-untyped-def (36): tip anotasyonu (-> None / -> Type / param: Any)
   - no-any-return (31): cast(T, ...) (no-op) veya float(...) sarma
   - union-attr/attr-defined: `assert x is not None` (mevcut invariant'ı belgeler)
   - arg-type: Path(...) sarma (str→Path, doğru); index/assignment: anotasyon genişletme
   - misc (conditional base class FPDF): # type: ignore[misc]
   3 paralel agent (disjoint dosya setleri) + serialization/pdf_report/graph_viz elle.

2) schema-drift + competition-compliance: PREDICTION_COLUMNS = 12 kolon
   (10 standart + OOD_Score/OOD_Flag — kasıtlı; ood_detector.pkl mevcut; build_prediction_frame
   zaten 12 üretiyor). ci.yml guard'ları (EXPECTED + len==10) 12'ye hizalandı.
   Resmi şartname çıktı-kolon formatı BELİRTMİYOR (araştırıldı: Tier-1 PDF okunamadı,
   Tier-3 "belirtilmemiş") → UNVERIFIED; §7.3 yalnızca F1.

3) U-009: jüri submission kolon-formatı UNVERIFIED kaydı (Google Groups'tan teyit edilmeli).
4) scripts/gen_readme_badges.py ruff-format (pre-existing).

Doğrulama: mypy 0 · ruff check temiz · ruff format --check 197/197 · consistency-gate 7/7 · compileall temiz.

+++ src/features/categorical_bio_features.py
@@ -88,7 +88,7 @@ def get_grantham_score(ref: str, alt: str) -> float:
     c1, p1, v1 = _GRANTHAM_CPV[r]
     c2, p2, v2 = _GRANTHAM_CPV[a]
     raw = (_GRANTHAM_ALPHA * (c1 - c2) ** 2 + _GRANTHAM_BETA * (p1 - p2) ** 2 + _GRANTHAM_GAMMA * (v1 - v2) ** 2) ** 0.5
-    return raw * _GRANTHAM_SCALE
+    return float(raw * _GRANTHAM_SCALE)
 
 
 GRANTHAM_PROPS = _GRANTHAM_CPV  # geriye dönük alias
